+++ estructural-patterns/src/main/java/br/com/cod3r/proxy/bank/services/BankOperations.java
@@ -1,7 +1,7 @@
 package br.com.cod3r.proxy.bank.services;
 
 public interface BankOperations {
-	void deposit(Long account, Long amount);
-	void withdraw(Long account, String passwd, Long amount);
+	void deposit(Long account, double amount);
+	void withdraw(Long account, String passwd, double amount);
 	void changePassword(Long account, String oldPassword, String newPassword);
 }
